--- src/component/test/test.jsx
@@ -1,8 +1,0 @@
-import { FavList } from "../client/jsx/FavList";
-export const Test = () => {
-  return (
-    <div>
-      <FavList />
-    </div>
-  );
-};
